Refactor model config actions
Move model config actions to separate file
