+++ lib/model-handler.js
@@ -8,39 +8,6 @@ const ModelMethod = require('./datamodel/model-method');
 const mixin = require('./util/mixin');
 const fsUtility = require('./util/file-utility');
 
-module.exports.findAllModels = function(workspace, cb) {
-  workspaceHandler.getFileList(workspace, function(err, files) {
-    if (err) return cb(err);
-    const modelFilePaths = files.Models || [];
-    const taskList = [];
-    const erroredFiles = [];
-    modelFilePaths.forEach(function(filePath) {
-      taskList.push(function(next) {
-        workspaceHandler.loadModelDefinition(workspace, filePath,
-        function(err) {
-          if (err)
-            erroredFiles.push({file: filePath, error: err});
-          next();
-        });
-      });
-    });
-    function callback(err) {
-      if (err) return cb(err);
-      let results = [];
-      const models = workspace.getAllModels();
-      if (models) {
-        Object.keys(models).forEach(function(key) {
-          let model = models[key];
-          results.push(model.getDefinition());
-        });
-      }
-      results = results.concat(erroredFiles);
-      cb(null, results);
-    }
-    workspace.execute(taskList, callback);
-  });
-};
-
 class ModelActions {
   create(cb) {
     const workspace = this.getWorkspace();
@@ -89,54 +56,35 @@ class ModelActions {
 
 mixin(Model.prototype, ModelActions.prototype);
 
-class ModelConfigActions {
-  create(modelId, facetName, cb) {
-    const workspace = this.getWorkspace();
-    const facet = workspace.getFacet(facetName);
-    const modelConfig = this;
-    facet.addModelConfig(modelConfig);
-    fsUtility.writeModelConfig(facet, cb);
-  }
-  update(facet, modelId, attrs, cb) {
-    const workspace = this.getWorkspace();
-    const modelConfig = this;
-    modelConfig.set(attrs);
-    const tasks = [];
-    tasks.push(function(next) {
-      fsUtility.writeModelConfig(facet, next);
+module.exports.findAllModels = function(workspace, cb) {
+  workspaceHandler.getFileList(workspace, function(err, files) {
+    if (err) return cb(err);
+    const modelFilePaths = files.Models || [];
+    const taskList = [];
+    const erroredFiles = [];
+    modelFilePaths.forEach(function(filePath) {
+      taskList.push(function(next) {
+        workspaceHandler.loadModelDefinition(workspace, filePath,
+        function(err) {
+          if (err)
+            erroredFiles.push({file: filePath, error: err});
+          next();
+        });
+      });
     });
-    workspace.execute(tasks, cb);
-  }
-}
-
-mixin(ModelConfig.prototype, ModelConfigActions.prototype);
-
-class ModelPropertyActions {
-  create(modelId, cb) {
-    const workspace = this.getWorkspace();
-    const model = workspace.getModel(modelId);
-    model.setProperty(this);
-    const tasks = [];
-    tasks.push(function(next) {
-      fsUtility.writeModel(model, next);
-    });
-    workspace.execute(tasks, cb);
-  }
-}
-
-mixin(ModelProperty.prototype, ModelPropertyActions.prototype);
-
-class ModelMethodActions {
-  create(modelId, cb) {
-    const workspace = this.getWorkspace();
-    const model = workspace.getModel(modelId);
-    model.setMethod(this);
-    const tasks = [];
-    tasks.push(function(next) {
-      fsUtility.writeModel(model, next);
-    });
-    workspace.execute(tasks, cb);
-  }
-}
-
-mixin(ModelMethod.prototype, ModelMethodActions.prototype);
+    function callback(err) {
+      if (err) return cb(err);
+      let results = [];
+      const models = workspace.getAllModels();
+      if (models) {
+        Object.keys(models).forEach(function(key) {
+          let model = models[key];
+          results.push(model.getDefinition());
+        });
+      }
+      results = results.concat(erroredFiles);
+      cb(null, results);
+    }
+    workspace.execute(taskList, callback);
+  });
+};
