+++ lib/workspace-manager.js
@@ -6,6 +6,8 @@ const facetHandler = require('./facet-handler');
 const dataSourceHandler = require('./data-source-handler');
 const middlewareHandler = require('./middleware-handler');
 const modelHandler = require('./model-handler');
+const modelConfigHandler = require('./model-config-handler');
+const modelPropertyHandler = require('./model-property-handler');
 
 /**
  * @class WorkspaceManager
